+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Journal des modifications
 - Accès aux rapports CSI et mi-parcours même lorsque la thèse est abandonnée
 - Module Soutenance : message informant des possibles présidents du jury (rang A) dans l'encart Association d'acteur SYGAL aux membres de jury. 
 - [FIX] Warning: Undefined variable $url in module/Depot/view/depot/fichier-these/deposes.phtml
+- [FIX] Recherche textuelle d'individu selon le type : jointure avec doctorant/acteur au cas où individu.type est null. 
 
 9.3.0
 -----